Add spinner in local installer longer steps
