+++ crates/fluvio-cluster/Cargo.toml
@@ -60,6 +60,7 @@ k8-client = { version = "5.0.0" }
 k8-metadata-client = { version = "3.0.0" }
 k8-types = { version = "0.2.0", features = ["app"]}
 async-channel = "1.6.1"
+indicatif = "0.16.2"
 
 [dev-dependencies]
 fluvio-future = { version = "0.3.7", features = ["task","fixture"] }
